+++ pyanalyze/name_check_visitor.py
@@ -937,6 +937,7 @@ class NameCheckVisitor(node_visitor.ReplacingNodeVisitor):
     current_enum_members: Optional[Dict[object, str]]
     _name_node_to_statement: Optional[Dict[ast.AST, Optional[ast.AST]]]
     import_name_to_node: Dict[str, Union[ast.Import, ast.ImportFrom]]
+    expected_return_value: Optional[Value]
 
     def __init__(
         self,
@@ -1797,8 +1798,10 @@ class NameCheckVisitor(node_visitor.ReplacingNodeVisitor):
         params: Sequence[SigParameter],
     ) -> FunctionResult:
         # Ignore generators for now.
-        if isinstance(return_set, AnyValue) or (
-            self.is_generator and info.async_kind is not AsyncFunctionKind.normal
+        if (
+            isinstance(return_set, AnyValue)
+            or return_set is NO_RETURN_VALUE
+            or (self.is_generator and info.async_kind is not AsyncFunctionKind.normal)
         ):
             has_return = True
         elif return_set is UNINITIALIZED_VALUE:
@@ -3089,11 +3092,12 @@ class NameCheckVisitor(node_visitor.ReplacingNodeVisitor):
             value = KnownNone
         else:
             value = self.visit(node.value)
-        if value is NO_RETURN_VALUE:
-            return
         self.return_values.append(value)
         self._set_name_in_scope(LEAVES_SCOPE, node, AnyValue(AnySource.marker))
-        if self.expected_return_value is NO_RETURN_VALUE:
+        if (
+            self.expected_return_value is NO_RETURN_VALUE
+            and value is not NO_RETURN_VALUE
+        ):
             self._show_error_if_checking(
                 node, error_code=ErrorCode.no_return_may_return
             )
@@ -3111,7 +3115,11 @@ class NameCheckVisitor(node_visitor.ReplacingNodeVisitor):
                     error_code=ErrorCode.incompatible_return_value,
                     detail=tv_map.display(),
                 )
-        if self.expected_return_value == KnownNone and value != KnownNone:
+        if (
+            self.expected_return_value == KnownNone
+            and value != KnownNone
+            and value is not NO_RETURN_VALUE
+        ):
             self._show_error_if_checking(
                 node,
                 "Function declared as returning None may not return a value",
@@ -3907,7 +3915,7 @@ class NameCheckVisitor(node_visitor.ReplacingNodeVisitor):
             if vals:
                 return unite_values(*vals)
             else:
-                return UNINITIALIZED_VALUE
+                return NO_RETURN_VALUE
         return local_value
 
     def visit_Attribute(self, node: ast.Attribute) -> Value: