+++ pyanalyze/stacked_scopes.py
@@ -368,7 +368,7 @@ class Constraint(AbstractConstraint):
                         yield TypedValue(self.value)
                     # TODO: Technically here we should infer an intersection type:
                     # a type that is a subclass of both types. In practice currently
-                    # _constrain_values() will eventually return AnyValue.
+                    # _constrain_value() will eventually return NoReturn.
                 else:
                     if not safe_issubclass(inner_value.typ, self.value):
                         yield value
@@ -1453,20 +1453,21 @@ def uniq_chain(iterables: Iterable[Iterable[T]]) -> List[T]:
 
 
 def _constrain_value(
-    values: Iterable[Value],
+    values: Sequence[Value],
     constraints: Iterable[Constraint],
     *,
     fallback_value: Optional[Value] = None,
     simplification_limit: Optional[int] = None,
 ) -> Value:
     # Flatten MultiValuedValue so that we can apply constraints.
-    values = [val for val_or_mvv in values for val in flatten_values(val_or_mvv)]
     if not values and fallback_value is not None:
         values = list(flatten_values(fallback_value))
+    else:
+        values = [val for val_or_mvv in values for val in flatten_values(val_or_mvv)]
     for constraint in constraints:
         values = list(constraint.apply_to_values(values))
     if not values:
-        return AnyValue(AnySource.unreachable)
+        return NO_RETURN_VALUE
     if simplification_limit is not None:
         return unite_and_simplify(*values, limit=simplification_limit)
     return unite_values(*values)